The variable type of the first input variable of TMath::Sort() function in QwEventBuffer.cc function was changed from type:unsigned to type:int. This was done inorder for the QwAnalysis code to be compatible with the current version of root,5-22-00. It was originally compiled using root v-5-12.
git-svn-id: https://qweaksvn.jlab.org/repos/QwAnalysis/trunk@121 b9700a98-9937-0410-b777-f318d6eb5bff

+++ Analysis/src/QwEventBuffer.cc
@@ -459,7 +459,7 @@ Bool_t QwEventBuffer::DataFileIsSegmented()
       local_index.resize(tmp_segments.size(),0);
       /*  Get the list of segments sorted numerically in        *
        *  increasing order.                                     */
-      TMath::Sort(tmp_segments.size(),&(tmp_segments[0]),&(local_index[0]),
+      TMath::Sort(static_cast<int>(tmp_segments.size()),&(tmp_segments[0]),&(local_index[0]),
                   kFALSE);
       /*  Put the segments into numerical order in              *
        *  fRunSegments.                                         */
